requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap5==2025.4
cssselect2==0.8.0
Django==4.2.20
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-environ==0.12.0
django-jazzmin==3.0.1
fontawesomefree==6.6.0
fonttools==4.57.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
weasyprint==65.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
